apps/will-do-api/pyproject.toml
[tool.poetry]
name = "will-do-api"
version = "0.1.0"
description = "To do API"
authors = ["William Phelps <williamphelps13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "0.6.0"
anyio = "4.2.0"
bcrypt = "4.1.2"
certifi = "2024.2.2"
cffi = "1.16.0"
click = "8.1.7"
cryptography = "42.0.2"
ecdsa = "0.18.0"
fastapi = "0.109.2"
h11 = "0.14.0"
httpcore = "1.0.2"
httptools = "0.6.1"
httpx = "0.26.0"
idna = "3.6"
packaging = "23.2"
passlib = "1.7.4"
pluggy = "1.4.0"
psycopg2-binary = "2.9.9"
pyasn1 = "0.5.1"
pycparser = "2.21"
pydantic = "2.6.1"
pydantic_core = "2.16.2"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
python-multipart = "0.0.7"
PyYAML = "6.0.1"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
SQLAlchemy = "2.0.25"
starlette = "0.36.3"
typing_extensions = "4.9.0"
uvicorn = "0.27.0.post1"
uvloop = "0.19.0"
watchfiles = "0.21.0"
websockets = "12.0"

[tool.poetry.group.dev.dependencies]
iniconfig = "2.0.0"
pytest = "7.4.4"
pytest-asyncio = "0.23.4"
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]
line-length = 88
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-line-length = "dynamic"
